+++ src/Organizations/OrganizationForm/OrganizationInterfacesForm/OrganizationInterfacesForm.js
@@ -4,7 +4,7 @@ import { FieldArray } from 'redux-form';
 
 import {
   Col,
-  Icon,
+  Loading,
   Row,
 } from '@folio/stripes/components';
 import { stripesConnect } from '@folio/stripes/core';
@@ -13,24 +13,26 @@ import { batchFetch } from '@folio/stripes-acq-components';
 import { interfacesResource } from '../../../common/resources';
 import OrganizationInterfacesList from './OrganizationInterfacesList';
 
-function OrganizationInterfacesForm({ orgId, mutator, storedInterfaces }) {
+function OrganizationInterfacesForm({ open, orgId, mutator, storedInterfaces }) {
   const [interfaces, setInterfaces] = useState();
+  const [isLoading, setIsLoading] = useState(true);
+
   const refreshInterfaces = useCallback(interfaceIds => {
+    setIsLoading(true);
     batchFetch(mutator.interfacesManualFetch, interfaceIds || [])
-      .then(setInterfaces);
+      .then(setInterfaces)
+      .catch(() => {
+        setInterfaces([]);
+      })
+      .finally(() => setIsLoading(false));
   // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   useEffect(() => refreshInterfaces(storedInterfaces), [refreshInterfaces, storedInterfaces]);
 
-  const isLoading = !interfaces;
-
-  if (isLoading) {
+  if (isLoading || !open) {
     return (
-      <Icon
-        icon="spinner-ellipsis"
-        width="100px"
-      />
+      <Loading />
     );
   }
 
@@ -67,12 +69,9 @@ OrganizationInterfacesForm.manifest = Object.freeze({
 
 OrganizationInterfacesForm.propTypes = {
   mutator: PropTypes.object.isRequired,
+  open: PropTypes.bool,
   orgId: PropTypes.string,
   storedInterfaces: PropTypes.arrayOf(PropTypes.string),
-};
-
-OrganizationInterfacesForm.defaultProps = {
-  storedInterfaces: [],
 };
 
 export default stripesConnect(OrganizationInterfacesForm);